+++ Changelog.md
@@ -9,6 +9,7 @@
 - Display message to player if the "Done" (end turn) button can't find your combatant.
 - Only allow the **UP** button to appear when the current Drama Card shows UP for the actor's faction.
 - Ensure correct state of "play card" button state when hero is added to/removed from a combat.
+- Detect when an **Approved Action** has been performed - add a note to the Chat Card in such cases.
 
 ### Bug Fixes